feat: move to checking against last 10 in database
This would break backwards compatibility if anybody else was using abusing this system. So counting it as a release. This is what I should be doing to help avoid skipping posts anyway.
It might mean we see a spike in some ancient posts getting made but should be okay since the scraper isn't still using the lid in previous and not the list